% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headlines}
\alias{headlines}
\title{NYT headlines by economic policy uncertainty}
\format{
An object of class \code{data.frame} with 11934 rows and 2 columns.
}
\usage{
headlines
}
\description{
NYT headlines by economic policy uncertainty
}
\keyword{datasets}
