% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig.R
\name{wig}
\alias{wig}
\title{WIG model training}
\usage{
wig(
  dates,
  docs,
  spec = wig_spec(),
  group_time = "months",
  svd_method = c("topics", "docs"),
  standardize = TRUE,
  verbose = FALSE
)
}
\description{
WIG model training
}
